+++ client-elastic4s/src/main/scala/com/klibisz/elastiknn/client/package.scala
@@ -1,0 +1,7 @@
+package com.klibisz.elastiknn
+
+import scala.concurrent.Future
+
+package object client {
+  type ElastiknnFutureClient = ElastiknnClient[Future]
+}