+++ examples/demo/webapp/app/views/utils.scala
@@ -1,0 +1,12 @@
+package views
+
+import models.SearchResult
+
+object utils {
+
+  def resultIsCorrect(result: SearchResult, exactResults: Seq[SearchResult]): Boolean =
+    exactResults.exists(_.id == result.id)
+
+  def durationSeconds(durMillis: Long): String = f"${durMillis / 1000f}%.4f seconds"
+
+}